Take advantage of Go 1.5 optimized zeroing.
Closes #286.
