+++ internal/zero/doc.go
@@ -1,0 +1,3 @@
+// Package zero contains functions to clear data from byte slices and
+// multi-precision integers.
+package zero
